+++ secfsdstools/c_download/secdownloading_process.py
@@ -16,7 +16,6 @@ class SecDownloadingProcess(BaseDownloadingProcess):
     """
         Downloading the quarterly zip files of the financial statement data sets
     """
-    FIN_STAT_DATASET_URL = 'https://www.sec.gov/dera/data/financial-statement-data-sets.html'
     FIN_STAT_DATASET_ARCHIVE_URL = \
         'https://www.sec.gov/data-research/sec-markets-data/financial-statement-data-sets-archive'
 
@@ -50,21 +49,8 @@ class SecDownloadingProcess(BaseDownloadingProcess):
             archive_hrefs = self.href_re.findall(archive_first_table)
             archive_hrefs = [f'https://www.sec.gov{href[6:-1]}' for href in archive_hrefs]
 
-        # reading data from the main url - starting with 2024q4.zip
-        LOGGER.info("reading table in main page: %s", self.FIN_STAT_DATASET_URL)
-        main_content = self.urldownloader.get_url_content(self.FIN_STAT_DATASET_URL)
-        main_tables = self.table_re.findall(main_content.text)
 
-        main_hrefs: List[str] = []
-
-        if len(main_tables) == 0:
-            LOGGER.warning("No table found at: %s", self.FIN_STAT_DATASET_URL)
-        else:
-            main_first_table = main_tables[0]
-            main_hrefs = self.href_re.findall(main_first_table)
-            main_hrefs = [f'https://www.sec.gov{href[6:-1]}' for href in main_hrefs]
-
-        hrefs = archive_hrefs + main_hrefs
+        hrefs = archive_hrefs
         return_value: List[Tuple[str, str]] = [(os.path.basename(href), href) for href in hrefs]
 
         return_value = [(n.replace("-archive", ""), p) for n, p in return_value]